feat: move tests into new directory
