+++ backend/src/tests/integrate/region.test.ts
@@ -1,5 +1,5 @@
 import request from 'supertest';
-import app from '../app';
+import app from '../../app';
 
 describe('POST /api/v1/region', () => {
   describe('add a new category name', () => {
